#57 [FEAT] 장소 검색 api 로직 작성

+++ app/src/main/java/com/kpaas/plog/data/dto/response/ResponseSearchDto.kt
@@ -4,14 +4,14 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponseSearchDto (
+data class ResponseSearchDto(
     @SerialName("searchedPlace") val searchedPlace: List<Place>
 )
 
 @Serializable
 data class Place(
-    @SerialName("latitude") val latitude: Long,
-    @SerialName("longitude") val longitude: Float,
+    @SerialName("latitude") val latitude: Double,
+    @SerialName("longitude") val longitude: Double,
     @SerialName("roadAddr") val roadAddr: String,
     @SerialName("placeInfo") val placeInfo: String
 )
